function_calling_MCP/realtime-master/pyproject.toml
[project]
name = "realtime-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13.1"
dependencies = [
    "audioop-lts>=0.2.1",
    "flask>=3.1.1",
    "getchar>=2.0.0",
    "googlesearch-python>=1.3.0",
    "mcp[cli]>=1.6.2",
    "numpy>=2.2.5",
    "openai[realtime]>=1.78.1",
    "pyaudio>=0.2.14",
    "pydub>=0.25.1",
    "python-dotenv[cli]>=1.1.0",
    "rich>=14.0.0",
    "sounddevice>=0.5.1",
]
